Create Ecosite / Ecosite Note XLSX NASIS Import Files

Create Ecosite / Ecosite Note XLSX NASIS Import Files

create_ESD_ecosites_import(file, coiids, ecositeids)

create_ESD_notes_import(file, coiids, author, notes)

create_note_from_ESD_ecosites(file, template, ..., sheet = "ESDList")

Arguments

file

output file name (either .XLSX or .CSV)

coiids

vector of component IDs

ecositeids

vector of ecological site IDs

author

author of note

notes

note content

template

a sprintf()-style format string up to 8192 bytes in length

...

values to be passed into template. Only logical, integer, real and character vectors are supported.

sheet

Default XLSX sheet name "ESDList"

Value

writes XLSX or CSV file

Examples

create_ESD_ecosites_import("test_esd.xlsx", 2770865, "F018XI205CA")
#> Loading required namespace: openxlsx

esdnotes <- create_note_from_ESD_ecosites("test_esd.xlsx", "Assigned %s %s")$note
#> Loading required namespace: odbc
#> Error : nanodbc/nanodbc.cpp:1021: 00000: [unixODBC][Driver Manager]Data source name not found and no default driver specified 
#> Warning: no local NASIS database available
#> Error in soilDB::dbQueryNASIS(soilDB::NASIS(), paste0("SELECT DISTINCT ecositeid, ecositenm FROM ecologicalsite WHERE ecositeid IN ",     soilDB::format_SQL_in_statement(unique(x)))): Failed to connect to NASIS database!

esdnotes
#> Error in eval(expr, envir, enclos): object 'esdnotes' not found

create_ESD_notes_import("test_esd_note.xlsx",
                        coiids = 2770865,
                        author = "Andrew Brown",
                        notes = esdnotes)
#> Error in data.frame(coiid = coiids, author = author, note = notes): object 'esdnotes' not found